Add block explorer to net up --kubernetes (#4582)
## Motivation

We're working on adding the Indexer/Explorer/Exporter into our
deployments.

## Proposal

This adds them to `linera net up --kubernetes` as a starting point.
Next, adding to the actual GCP deployment.

## Test Plan

Started a local k8s network, sent some transactions to it, and managed
to see the information in the block explorer:

![Screenshot 2025-09-16 at
12.18.34.png](https://app.graphite.dev/user-attachments/assets/80cd3746-4f75-4d0b-aa5d-1daa900be025.png)

![Screenshot 2025-09-16 at
12.18.55.png](https://app.graphite.dev/user-attachments/assets/b765f25d-271e-47ad-987a-7691c5b5c755.png)

## Release Plan

- Nothing to do / These changes follow the usual release cycle.