Restructure the `RemoteNetTestConfig` type. (#4684)
We have some errors "ERROR linera: Error is Failed to close chain" in
the running of `RemoteNet` tests, either against a local set of
validators or the TestNet. Those errors are not critical, but are
annoying.

The following has been done:
* The `RemoteNetTestingConfig` takes some argument `faucet_url:
Option<String>`. However, that argument is never used. Instead, the
faucet URL is set from the environment variable `LINERA_FAUCET_URL`.
This is the case for remote-net tests and for the test against an
existing testnet.
* Therefore, the `faucet_url` argument can be dropped.
* Instead the argument `close_chain` is added. If false, the chain
leaks.

The net effect of this PR is to remove the error from the running of the
remote-net tests.
What it does not:
* It does not change the close-chain feature, whether we want to
continue it or not.
* It leaves the chains for all other kinds of tests.
* It does not address the permission problem that prevents the closure
for some chains in `remote-net`.
* If a chain were to fail closing, no error is reported.