Introduce the `multi_get` for `MapView` and make the `fn entries` used it. (#4634)
## Motivation

The `fn entries` should be using a `multi_get`, but the function was
missing.

## Proposal

It was introduced straightforwardly, by taking `CollectionView` /
`ReentrantCollectionView` as a model.

## Test Plan

The unit test have been added for the functionality.

## Release Plan

It is fully compatible with TestNet / DevNet and could impact the GoL
which is using `MapView`.

## Links

None.