Adding some faucet metrics (#4692)
## Motivation

The current metrics we have on the faucet its stuff that we're exporting
in the client code, but we don't have any faucet specific metrics.

## Proposal

Add some faucet specific metrics.

## Test Plan

Deploy a network with this code, see the faucet metrics being exported.
Used this to create a dashboard as well with these metrics:

![Screenshot 2025-09-30 at
01.01.20.png](https://app.graphite.dev/user-attachments/assets/ffb1efc3-117a-45c6-8153-4031724d569a.png)

## Release Plan

- Nothing to do / These changes follow the usual release cycle.