Rename `authenticated_signer` as `authenticated_owner`. (#4696)
## Motivation

The term "authenticated signer" is not great, since when a message is
transmitted, the term is used while who is going to sign is going to be
another user/owner.

## Proposal

Rename the "authenticated signer" as "authenticated_owner".

## Test Plan

The CI seems to be the right way.

## Release Plan

It is likely to be a breaking change.

## Links

None.