`linera-service`: faucet URL env var (#4710)
## Motivation

Quality of life: it's pretty annoying when using `linera net up` to have
to repeat the faucet URL all the time.

## Proposal

Support setting the faucet URL through an environment variable for the
CLI, akin to `LINERA_WALLET` and friends. This reduces the need for
repeating the URL a lot when testing on a local network.

## Test Plan

Used locally.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)