Improve a few error and log message strings. (#4734)
## Motivation

The pending block clearing message is INFO-level, but it's spammy
because it also shows for each successfully committed block.

The `UnexpectedBlockHeight` error is confusing, because it's not clear
which height is which.

## Proposal

Make the pending block message DEBUG.

Change the message of `UnexpectedBlockHeight`.

## Test Plan

CI

## Release Plan

- We could backport it to `testnet_conway`, then:
    - release a new SDK,
    - do a validator hotfix.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)