Improve a few error and log message strings. (#4734) (#4735)
Backport of #4734.

## Motivation

The pending block clearing message is INFO-level, but it's spammy
because it also shows for each successfully committed block.

The `UnexpectedBlockHeight` error is confusing, because it's not clear
which height is which.

## Proposal

Make the pending block message DEBUG.

Change the message of `UnexpectedBlockHeight`.

## Test Plan

CI

## Release Plan

- Not urgent, but will be included in:
    - next SDK,
    - next validator hotfix.

## Links

- PR to main: #4734
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)