Web tests using Vitest and Playwright (#4704)
## Motivation

Currently all our testing for the Web client is manual. We'd like to
know when PRs break our example apps.

## Proposal

This PR sets up conventions for our Web tests that enable us to test in
a variety of different environments. Going forward we can add more and
more in-depth tests to cover more scenarios, and if we backport this PR
we can also use these tests as an automated healthcheck for the deployed
Web apps.

We unify the Web tests using [Vitest](https://vitest.dev/) with a
[Playwright](https://playwright.dev/) backend to run tests in real
browsers (currently just Chromium, but we should also test at least on
WebKit and Firefox once
https://github.com/linera-io/linera-protocol/issues/4235 is fixed).

Add a couple of basic tests for `@linera/client`: getting a wallet from
the faucet and initializing the client.

## Test Plan

This is the test plan :)

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)