Add support for chrome trace exporter (#4712)
## Motivation

Right now for tracing information we can only visualize it using Tempo,
which doesn't really work for local client runs.

## Proposal

Add support for `ChromeTraceExporter`, which generates a JSON file that
can be ingested by the Perfetto UI for visualization.

## Test Plan

Ran the client locally with the `OTEL_EXPORTER_TYPE=chrome_json`, got
the JSON file and managed to see the tracing data in the Perfetto UI

https://ui.perfetto.dev/#!/viewer?local_cache_key=00000000-0000-0000-f5e0-334428275111

## Release Plan

- Nothing to do / These changes follow the usual release cycle.