Address a bug in `try_load_entries` of `CollectionView`. (#4753)
## Motivation

In `CollectionView`, we have `KeyTag::Index` / `KeyTag::Subview`, but
they were wrongly
assigned in the function `try_load_entries`.

## Proposal

The correction was straightforward, once it was found.

## Test Plan

A random test has been added. This test concerns both the
`CollectionView` and `ReentrantCollectionView`.
This test covers both the found bug, the preceding one with
`W::NUM_INIT_KEYS == 0`. Therefore, the two
added tests in the `view_test.rs` are removed.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

A test case was created in
[linera_end_to_end_tests/graphql_induced_crash](https://github.com/MathieuDutSik/linera_end_to_end_tests).
It could be put as
an additional integration test that exercises the GraphQL capabilities
of Linera.