Implement the caching of the `find_{keys,key_value}_by_prefix` operations. (#4576)
## Motivation

The `find_keys_by_prefix` has emerged as a choke point for some
benchmarks.
Considering that it is the only part that is not cached, it is
interesting to consider this.

## Proposal

The implementation is done in the following way:
* The `find_{keys,key_values}_by_prefix` are cached.
* The `delete_prefix` leads to a `find_key_values_by_prefix` entry in
the cache.
* All the information about the operation is used to avoid making some
calls to the storage.

## Test Plan

The CI.
Some unit tests have been added to the `lru_caching.rs` file.

## Release Plan

The input file would change because some additional entries have been
added.
However, this can still be ported to existing DevNet / TestNet.

## Links

None.