Use `multi_get` for preprocessed_blocks. (#4788)
## Motivation

The function `block_hashes` has emerged as a limiting factor for the
running of the system.
An iteration of `get` operations is ripe for optimization.

## Proposal

We use a `multi_get`. Since what is downloaded is just a hash, there is
no risk of memory problems.

## Test Plan

The CI.

## Release Plan

This kind of acceleration can be backported to TestNet Conway easily.

## Links

None.