Improve test coverage of `set_view.rs` from 38.18% to 86.78% (#4782)
Test coverage helps to strenghten the quality of the software. and allow
to know precisely what is working or not.

This was done by systematically adding tests.
In passing two typos from previous PRs were cleaned up.

Testing is the purpose of the PR.

This could be backported to TestNet Conway.

None.