Improve test coverage of `set_view.rs` from 38.18% to 86.78% (#4782)
## Motivation

Test coverage helps to strenghten the quality of the software. and allow
to know precisely what is working or not.

## Proposal

This was done by systematically adding tests.
In passing two typos from previous PRs were cleaned up.

## Test Plan

Testing is the purpose of the PR.

## Release Plan

This could be backported to TestNet Conway.

## Links

None.