Add `try_load_all_entries` to `CollectionView`. (#4763)
## Motivation

Just like for `ReentrantCollectionView`, we can load all entries at once
in `CollectionView`.

## Proposal

The code was added in a straightforward way.

## Test Plan

The test has been to the CI. The consequence is that the
`try_load_all_entries` has been tested for `ReentrantCollectionView`.
This led to the removal of the `Q` types that was not used.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.