Backport: Check the block height before the epoch in `process_timeout` (#4811) (#4813)
## Motivation

When processing a timeout, we check whether the epoch of the certificate
is correct before checking whether the height is correct - this may lead
to spurious `InvalidEpoch` errors.

## Proposal

Validate the height first.

This is the backport of #4811 to the testnet branch.

## Test Plan

CI

## Release Plan

- Nothing to do

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)