Avoid clone operation in `read_multi_values_internal`. (#4820)
## Motivation

The function `read_multi_values_internal` has to deal with the fact that
keys can repeat. So a HashMap is built for storing the key indices.
Unfortunately, this leads to a clone operation, while we could have
avoided the clone operation for the last one. This is especially bad
since in most cases, we would have just one entry.

## Proposal

Use the `split_last()` in the access to the indices.


## Test Plan

The CI.

## Release Plan

That can be backported to TestNet Conway.
And this can be a good idea since read operations are problematic on
TestNet Conway.

## Links

None.