Introduce `multi_get_pairs` operations and group another access to `preprocessed_blocks`. (#4827)
## Motivation

There are more uses of `multi_get` that are possible to group.
We also introduced some `multi_get_pairs` functions in order to avoid
some `.zip` constructions.

#4802 

## Proposal

The `multi_get_pairs` is introduced for `*MapView` and `LogView`.
For the `*CollectionView`, we introduce similar functions.

The pattern in `state.rs` was grouped. When possible the `.zip` were
eliminated.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.