Backport: Do not process inboxes of non-owned chains (#4825) (#4832)
## Motivation

A service running without skipping process-inbox would spam the logs
with a `warn` level message if it followed a non-owned chain.

## Proposal

Skip processing the inbox in the listener if the chain isn't owned.

This is a backport of #4825 to the `testnet_conway` branch.

## Test Plan

CI

## Release Plan

- Nothing to do

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)