Key the execution state cache with the block hash (#4842)
## Motivation

If a confirmed certificate was submitted to a validator with a correct
execution state hash, but with an incorrect execution outcome, if the
validator already had the execution state hash in its cache, it wouldn't
re-execute the certificate and notice anything wrong.

## Proposal

Use the block hash instead of the execution state hash as the key in
`execution_state_cache`. This way if the outcomes are wrong, the
validator will re-execute the block and notice mismatching outcomes.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- #4824 - a related PR on the `testnet_conway` branch that exposed the
issue.