Eliminate completely the "fn load" implementation from the native views (#4844)
## Motivation

The function "`fn load`" was of two different forms. Either pre-load,
read the values, and post-load. Or just post-load.

## Proposal

Implement the code directly in the trait.
The same work is not done for the derived code because the derive code
contains the metrics.

## Test Plan

The CI.

## Release Plan

Normal release cycle. Could be merged in TestNet Conway, but that is not
needed.

## Links

None.