Synchronize before applying changes to the committee. (#4854)
## Motivation

At the moment it would probably not commit anything, but in the future
it might become dangerous to use `change-validators` (or `set-validator`
or `remove-validator`) without first using `sync`: It would take an
outdated committee, apply the modifications, and then use the result as
the new committee.

## Proposal

Synchronize the admin chain before getting the committee.

## Test Plan

This is just a safeguard so that we don't introduce a footgun in the
future, if we modify the client to automatically and silently fetch the
missing blocks inside `execute_operations`.

## Release Plan

- These changes should be backported to `testnet_conway`, then
    - be released in a new SDK.

## Links

- Related: #4833
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)