Add `NotificationsExt` trait to await specific notifications. (#4834)
## Motivation

The test-only `eventually` function waits in a loop with increasing
delays until a condition is satisfied. But usually these conditions are
satisfied exactly when the node processes a certain block and outputs
the corresponding notification.

## Proposal

Add a test-only `NotificationsExt` trait to simplify waiting for a
particular notification. Use it instead of several `eventually` and
`sync` calls in the tests.

## Test Plan

Most tests were changed to use this instead of `eventually`. Only the
ones waiting for the faucet weren't, because it currently doesn't
support GraphQL subscriptions for notifications.

## Release Plan

- These changes _could_ be backported to `testnet_conway`, to keep the
diff small.

## Links

- Closes #4718.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)