Remove `load_entry_or_insert` from `CollectionView`. (#4836)
## Motivation

The function `load_entry_or_insert` loads an entry or inserts it and
then returns a reference that cannot be used.
The returned reference is not mutable, but the insertion feature forces
the function to take a `&mut self`. So,
we get the worst of both worlds:
* We need a mutable input.
* The returned object is not mutable. Also, if it has been created, then
it is a default, which is not very interesting.

## Proposal

The following steps are being made:
* Remove altogether the `load_entry_or_insert` function.
* Move the `do_load_entry_mut` code into `load_entry_mut`.
* The `load_entry_or_insert` was used only for test purposes. The code
is replaced by `load_entry_mut`.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.