Introduce reading of keys by blocks for DynamoDb. (#4823)
## Motivation

For `DynamoDB`, we are executing `read_multi_values` queries by creating
`N` queries for `N` access to the storage. One for each key. But it is
possible to group the queries so that we have fewer queries.


## Proposal

We use the `BatchGetItem` functionality of `DynamoDB`.
We cannot group all reads in one single query. But instead of sending
`N` queries, we send about `N/40`, which is a significant gain.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

It does not appear important to put it on TestNet Conway since DynamoDB
is not used for it.

## Links

None.