Address a bug with the `root_keys` in `linera-storage-service`. (#4858)
## Motivation

There is a bug in the handling of the root_keys in the
`linera-storage-service`. The `root_key` is directly appended to the
`start_key`. The result is an object which is not prefix-free.

## Proposal

Replace
```rust
start_key.extend(root_key)
```

by
```rust
start_key.extend(bcs::to_bytes(&root_key)?)
```

by doing so, we no longer have ambiguity in the construction of
`start_key`.


## Test Plan

The CI.
Added a test that fails on main on that precise problem.

## Release Plan

Since, the `linera-storage-service` is not used in TestNet Conway, no
backport is needed.

## Links

None.