Change the API of `list_root_keys`. (#4860)
## Motivation

The API of `list_root_keys` is taking the config and namespace. This is
a little bit illogical, since the function should be a function of the
database itself.

## Proposal

Restructure the API accordingly.

## Test Plan

The CI should catch the problems.

## Release Plan

This is an internal change, which is useful for debugging and
consistency.
It can be backported to TestNet Conway but that does not appear to be
essential.

## Links

None.