Correct a bug in `root_key` handling in `RocksDb`. (#4859)
## Motivation

The bug addressed in PR #4858 is also present in `RocksDb`.

## Proposal

The same solution is applied. The test code inserted in
`linera-storage-service` is added to `root_key_admin_test` so that all
storage are now tested on this possible issue.

## Test Plan

The CI.

## Release Plan

This PR should NOT be merged into TestNet Conway.

The reason is that while the bug being addressed is real, correcting it
means that we lose access to the
RocksDB storage, through the standard API.

Backport in TestNet Conway after main.

## Links

None.