Insert a test function for the storage and tests for Memory, ScyllaDb and DynamoDb. (#4821)
## Motivation

The `linera-storage` is a critical part of `linera`. However, we do not
have unit tests for the storage
functionality, even if it is very important. Instead, the tests are done
in `linera-views`, in the `end-to-end`
tests and other tests.

## Proposal

Introduce a function that takes a storage and does the tests.
Introduce three separate functions for the memory, DynamoDb, and
ScyllaDb.


## Test Plan

The PR introduces some tests.
The memory tests are immediate, the DynamoDb takes 0.3s, and the
ScyllaDb 0.69s.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
There is no need to put it on TestNet Conway.

## Links

None.