No fees for rejecting (#4875)
## Motivation

We charge fees for all outgoing messages, including bounced ones. That
means that a chain with very low balance can become unable to reject
tracked messages.

The client doesn't expect this, and keeps looping and setting the
failing message to `Reject`, even if it already is.

## Proposal

Don't charge for bouncing messages.

Also, charge only for the length of the operation outcome payload,
rather than the length of its serialized size, so that an empty outcome
is free, too.

Finally, we now break out of the loop in the client if a _rejected_
message fails.

## Test Plan

A new test was added that reproduced the issue.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- I will prepare a separate PR for a migration of `testnet_conway` to
this new execution semantics. We can decide later whether we want to do
that.

## Links

- Found in: https://github.com/linera-io/linera-protocol/issues/4874
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)