Extend MessagePolicy filtering to events (#4877)
## Motivation

We might want to process the inbox of a chain selectively in some cases
(eg. if a chain has a lot of incoming messages in the inbox and lacks
the funds to process them, we could send a transfer and only process the
messages from the sender of the transfer first).

## Proposal

~~Add a `--senders` argument to `process-inbox`, specifying the chains
from which we want to process messages. If this argument is not present,
process all incoming messages as before.~~

Turns out the feature already existed - only extend filtering by chain
to events. Also, test message filtering by chain.

## Test Plan

CI

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)