New BlockExecuted notification & support querying app state at intermediate execution states. (#4867)
Some applications may decide that it is safe to update their UIs
whenever the block proposal is handled locally (i.e. it's executed, set
as pending proposal and sent to validators) instead of waiting for the
confirmed block notification.

Add new notification `Reason::BlockExecuted { height, hash }` which is
emitted after staging the block locally (i.e. it's set as pending and
executed). The `hash` can be used to query the applications state even
before it's committed. To support this the `query_application`
functionality was extended to support querying at intermediate, cached,
execution state views. When querying, no hash is provided, we query the
committed state. Similarly, if the provided hash is no longer in the
cache (b/c it was already committted in the meantime or some other
proposal superseeded this one), the committed state is queried.

A test with notifications was extended to verify that the event is
emitted.

- These changes should
    - be released in a new SDK,
    - backported to Testnet.

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)