Backport: Extend MessagePolicy filtering to events (#4877) (#4879)
## Motivation

`MessagePolicy` can already be used to decide which messages we want to
receive. We would like to be able to control events as well.

## Proposal

Use the list of chains from `MessagePolicy` to restrict which chains'
events are processed as well.
This is a backport of #4877 

## Test Plan

CI

## Release Plan

- These changes should be released in a new SDK

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)