Add AssertStorage for `linera-proxy` and `linera-server`. (#4887)
## Motivation

When running `linera-proxy` / `linera-server`, we should check that the
storage has been migrated.

## Proposal

The wrong idea (that was attempted before) was to check in
`linera-storage`). This is the wrong place
since it does not have access to the context of the call.

Instead, the idea is to introduce an `IsStorageV1` in parallel to
`StorageMigration` and to test it in
`linera-proxy` / `linera-server` near to the main.rs.+

## Test Plan

The CI.

## Release Plan

To be merged in TestNet Conway.

## Links

None.